Reverb: Add a custom InstallCommand that overrides install_lib.
This tells setuptools that the package is not a "pure" Python-only package.
Normally when a package is not pure it is because it also compiles a Python
C Extension. However that is not the case here since we directly call bazel to compile
the shared libraries. This means that we need to manually set the install_lib to
so setuptools creates a "platform" wheel.

PiperOrigin-RevId: 309910999
Change-Id: I671e9852556f84e87dacaaa5140c743fb188c479