Treat UnavailableError as the only transient error type in the writer.
Previously we have been treating CancelledError and DeadlineExceededError as transient as well. This has likely been a large source of pain as this mean that if the server is shut down (not due to failure but due to stop being called) then the writer will continue to retry indefinetely. Note that this may still occur if the writer is not connected to the server at the time it is shut down but at least now the behavior is not ALWAYS bad. As for DeadlineExceeded, we don't set a deadline so there is no reason to expect it to ever be thrown.

PiperOrigin-RevId: 320567938
Change-Id: I4728d9abbf942897869f3038958f0479eb02f996