Reverb: Adding `Table::SampleFlexibleBatch`.
When sampling a batch of items we currently lock and unlock the table in between each sample. This CL introduces the option of greedily sample extra items without releasing the lock in between. That is, after the first sample has been approved by the rate limiter, check if more can be done without awaiting state changes and if so, go ahead and sample another item. This has the potential of reducing lock switching (and in theory improving CPU cache hits and reducing ping ponging).

NOTE: This CL doesn't change any behavior as the default batch sizes is set as 1.
PiperOrigin-RevId: 321749254
Change-Id: I8476edc98f2157ce25e8414ae67a57a347246d97