Add (optional) exact flow control to the `Writer`.
Items are currently sent to the server over a stream. The server will stop reading data from the buffer when blocked by the rate limiter which propagates to the client which is blocked from writing more data to the gRPC buffer. This works great when the data in each step is large enough to quickly fill up the gRPC buffer but can lead to unexpected behavior when the steps are small (i.e a large number of steps are placed on the gRPC buffer even though the rate limiter is blocking).

This CL introduces an optional parameter `max_in_flight_items` to the Writer which when set limits the number of items allowed to exist in the limbo between client and server. The syncronization is achieved by changing the stream to be bidirectional and making the server send responses for each item that has been inserted. These responses are read by the writer and optionally used to make the server and client alignment stricter.

PiperOrigin-RevId: 323556818
Change-Id: I60e8c5578ffb0a1423fdaa5e6458aaaa8778572b