Adding method `append_sequence` to `Writer`.
append_sequence allows multiple timesteps to be pushed to a writer with a single call. The writer will internally slice the tensor resulting in the same results as if append was called once for each "row" of the batch. Doing these manipulations in c++ can be meaningfully more efficient for large batches.

PiperOrigin-RevId: 330269184
Change-Id: I9cb9b8a2f463ae2dc707985e4745c04cc9b452f3