Bypass gRPC to sample from in process Table.
NOTE! This change was originally proposed and implemented by jmacglashan@gmail.com, piyushk@cogitai.com and declan@cogitai.com.

When the sampler is runnning in the same process as the Reverb server (e.g single process agent), the overhead imposed by gRPC can be avoided by accessing the underlying `Table` directly. This CL introduces a protocol that allows the server to safely transfer ownership of a shared_ptr<Table> to the client iff both are running in the same process. The `Sampler` has then been extended to make use of the direct access to the Table.

An ad hoc benchmark was used to evaluate sample throughput in a single process setting. We find that direct access result in 30% - 1000% higher max QPS depending on the setup. The impact is largest for very small payloads (<= 1kB). The results is outlined in details in this [colab](https://colab.research.google.com/drive/1rERA19XDEGkqJSOKr-CKS9z07a6Rnm1L?usp=sharing). The plots can also be found [here](https://imgur.com/a/bEw9nLQ).

PiperOrigin-RevId: 343109748
Change-Id: I063f57459157958e312b40394f62ca0b0ecda931