Modify `RecordioCheckpointer` and `TFrecordCheckpointer` to allow loading initial checkpoint from a different experiment.
In particular, the two classes have now an optional member `fallback_checkpoint_dir_`.
If a `fallback_checkpoint_dir` is passed to the constructor of `RecordioCheckpointer` or `TFrecordCheckpointer` and iff no checkpoint is found in the root directory, the checkpointers attempt to load a checkpoint from `fallback_checkpoint_dir`.
Note that by first attempting to load from the root directory and then only loading the fallback checkpoint iff the root directory is empty we are effectively using the fallback checkpoint as a way to initialise the service with a checkpoint generated by another experiment.

This change required also a small change in /l/d/r/control/minrl/backends:reverb_test. See in line comment for this.

PiperOrigin-RevId: 357185909
Change-Id: Id3eac4576aaafce10e5b5f32cd298ee7fe05a0c5