Replace zip with itertools.zip_longest in _tree_union.
zip stops when the shortest of the iterables is exhausted. This means that it is impossible
to add new fields when appending either tuples or lists. For example:

```
writer.append((0,))
writer.append((0, 1))
```

would raise a ValueError because self._structure wouldn't update to reflect the fact that
we should have 2 columns.

This is fixed by using itertools.zip_longest, which iterates until the longest iterable is
exhausted. Values for shorter iterables are replaced by `None`.

PiperOrigin-RevId: 361998620
Change-Id: I3cfd7b9bff53619589844c8ffa162a2c66d7e52b