Check to see if the number of elements in data matches the number of elements in self._structure before checking that they are the same structure.
`tree.assert_same_structure()` is a slow call and this check allows us to avoid calling it when data and self._structure are obviously different.

We could avoid the `tree.assert_same_structure()` call altogether. This means that we would deem two structures to be the same if all of their leaves have the  same paths (see `tree` for a description of what paths are).

PiperOrigin-RevId: 368178338
Change-Id: I9d99f77e0379f3f8eb5e35e91974576602ad2d1a