Allow for multiple `ChunkData` to be sent in the same `SampleStreamResponse`.
The `TrajectoryWriter` creates separate `ChunkData` for each column in the input data. This means that the number of `ChunkData` that needs to be sent when retrieving a sample is increased by a factor of `num_tensors_per_step`. Currently each `SampleStreamResponse` contains at most one `ChunkData` which means that the number of responses required to fulfill a sample request is scaled by the same number.

This CL changes `SampleStreamResponse` so that multiple `ChunkData` can be packed into the same message. An e2e benchmark were the data contained 8 columns shows a ~40% improvement (60k QPS -> 80k QPS), indicating that the change has the desired effect.

PiperOrigin-RevId: 372307569
Change-Id: I76b64bd6054b28aea4fc7f828e97ea55b06975df