Fix a bug where dataset iteration and rate_limiter_timeout_ms was broken.
Sampler workers run in separate threads from the dataset, meaning that they
execute sampling from the server in an async fashion from calls to get
the next value from the dataset.

What this means is that sampler workers may time out even though no data
is being requested via the dataset.  This is because the rate limiter timeout
time was set when the worker inner collection loop is executed (ans run for
e.g. typically 10000 iterations).  To fix this we need to do two things:

* Update the timeout after new data has come in and is put on the shared queue.
* Ignore the async timeouts on the worker if the dataset/client is not waiting
  for more data.

The latter part requires creating additional control flow where the sampler
"client" thread marks that it wants more data, during which the timeout is
respected.  Once it has received some items it marks that it no longer needs
more data.  In the implementation here, this control flow is done by the Queue:
while a user is Push or Popping, a counter is incremented.  The worker can check
to see if anyone is waiting to Pop when a timeout comes in.  If not, it can ignore
the timeout.

This version fixes the local sampler; a followup will fix the GRPC-based sampler.

PiperOrigin-RevId: 378689752
Change-Id: I3b286297efbaf91f200994251a464767bcc2499b