Add StreamingTrajectoryWriter.
The writer is similar to TrajectoryWriter but instead of optimizing reliability
and writer-latency, it optimizes memory usage on the writer-side. It streams all
chunks as soon as they are assembled in a synchronous matter. If streaming a chunk
fails, the entire episode is considered corrupt and no replay items can be created
for the rest of the episode.

PiperOrigin-RevId: 385780080
Change-Id: Ife15da75a8a0ca5d0f5f5ecc6edaf06e41d215d2