Stop waiting for confirmations from an unavailable server and instead retry properly.
Currently, if the server becomes unavailable after one or more items have been written to the stream but before the confirmation has been received the `TrajectoryWriter` will try to wait for the confirmations which is likely to cause deadlocks.

This CL fixes the behaviour so that unconfirmed items are resent to the (potentially new) server when a new connection is created.

Bonus:
 * Fix the count of items waiting in the queue in the deadline exceeded error message.
 * Deflaky the tests (now 100k runs passes).
PiperOrigin-RevId: 386215360
Change-Id: I03f1a7e7a5904e0927df4bd31411673048472540