Relax legacy `Writer` signature checking to be compatible with trajectory signatures.
Currently individual timesteps are checked against the Table signature, assuming that the signature describes a single timestep. However, the TrajectoryWriter requires that signatures are defined for the entire trajectory rather that a timestep. This makes it hard to manage codebases which uses both `Writer` and `TrajectoryWriter`.

This CL relaxes the signature check in the legacy writer slightly to make it compatible with signatures defined on the trajectory level.

PiperOrigin-RevId: 425322853
Change-Id: Ic7ee190ff520eadb447468a70f97da7bc2e58887