Early fail and more verbose checkpoint loading.
Currently the checkpoints are fully read from disk before the validity of the metadata content has been verified. This is a massive waste of time if the checkpoint ends up being corrupted. To avoid this we reverse the order in which data is loaded so that metadata is loaded and verified BEFORE the compressed trajectory data is loaded.

Furthermore, to make future debugging easier we also introduce more logging with this CL. This should be particularly helpful when it takes a very long time to load the trajectory data from disk.

PiperOrigin-RevId: 436525376
Change-Id: I022ec0325c1409380d46649e34133c6e47e41141